third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-lab-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property color value 'lab(calc(50 + (sign(1em - 10px) * 10)) 30 50 / 50%)'
  assert_equals: expected "lab(60 30 50 / 0.5)" but got "lab(40 30 50 / 0.5)"
[FAIL] Property color value 'oklab(calc(0.5 + (sign(1em - 10px) * 0.1)) 0.3 0.5 / 50%)'
  assert_equals: expected "oklab(0.6 0.3 0.5 / 0.5)" but got "oklab(0.4 0.3 0.5 / 0.5)"
[FAIL] Property color value 'lab(60 30 50 / calc(50% + (sign(1em - 10px) * 10%)))'
  assert_equals: expected "lab(60 30 50 / 0.6)" but got "lab(60 30 50 / 0.4)"
[FAIL] Property color value 'oklab(0.6 0.3 0.5 / calc(50% + (sign(1em - 10px) * 10%)))'
  assert_equals: expected "oklab(0.6 0.3 0.5 / 0.6)" but got "oklab(0.6 0.3 0.5 / 0.4)"
[FAIL] Property color value 'lch(calc(50 + (sign(1em - 10px) * 10)) 30 50deg / 50%)'
  assert_equals: expected "lch(60 30 50 / 0.5)" but got "lch(40 30 50 / 0.5)"
[FAIL] Property color value 'oklch(calc(0.5 + (sign(1em - 10px) * 0.1)) 0.3 50deg / 50%)'
  assert_equals: expected "oklch(0.6 0.3 50 / 0.5)" but got "oklch(0.4 0.3 50 / 0.5)"
[FAIL] Property color value 'lch(60 30 50deg / calc(50% + (sign(1em - 10px) * 10%)))'
  assert_equals: expected "lch(60 30 50 / 0.6)" but got "lch(60 30 50 / 0.4)"
[FAIL] Property color value 'oklch(0.6 0.3 50deg / calc(50% + (sign(1em - 10px) * 10%)))'
  assert_equals: expected "oklch(0.6 0.3 50 / 0.6)" but got "oklch(0.6 0.3 50 / 0.4)"
[FAIL] Property color value 'lab(calc(50 + (sign(2cqw - 10px) * 10)) 30 50 / 0.51)'
  assert_equals: expected "lab(60 30 50 / 0.51)" but got "lab(40 30 50 / 0.51)"
[FAIL] Property color value 'oklab(calc(0.5 + (sign(2cqw - 10px) * 0.1)) 0.3 0.5 / 0.51)'
  assert_equals: expected "oklab(0.6 0.3 0.5 / 0.51)" but got "oklab(0.4 0.3 0.5 / 0.51)"
[FAIL] Property color value 'lch(calc(50 + (sign(2cqw - 10px) * 10)) 30 50 / 0.51)'
  assert_equals: expected "lch(60 30 50 / 0.51)" but got "lch(40 30 50 / 0.51)"
[FAIL] Property color value 'oklch(calc(0.5 + (sign(2cqw - 10px) * 0.1)) 0.3 50 / 0.51)'
  assert_equals: expected "oklch(0.6 0.3 50 / 0.51)" but got "oklch(0.4 0.3 50 / 0.51)"
Harness: the test ran to completion.
